requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
discord.py==2.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pendulum==2.1.2
python-dateutil==2.8.2
pytzdata==2020.1
redis==4.5.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
yarl==1.8.2
